+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## Package: iqnection-silverstripe-modules-formbuilder
 # Change Log
 
+## 1.0.2
+- Improved duplication of forms
+- Added ability to resend submission emails
+
+
 ## 1.0.1
 - bug fixes and improvements in form duplication
 - bug fixes in payments